fix plain java example in ch 1
git-svn-id: http://camelinaction.googlecode.com/svn/trunk@271 2ef4527c-bcc6-11de-8784-f7ad4773dbab

+++ chapter1/file-copy/src/main/java/camelinaction/FileCopier.java
@@ -32,11 +32,13 @@ public class FileCopier {
         
         File[] files = inboxDirectory.listFiles();
         for (File source : files) {
-            File dest = new File(
-                    outboxDirectory.getPath() 
-                    + File.separator 
-                    + source.getName()); 
-            copyFile(source, dest);
+            if (source.isFile()) {
+                File dest = new File(
+                        outboxDirectory.getPath() 
+                        + File.separator 
+                        + source.getName()); 
+                copyFile(source, dest);
+            }
         }
     }
